requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
astroid==3.3.11
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.8.3
charset-normalizer==3.4.3
dill==0.4.0
flake8==7.3.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.1.0
mccabe==0.7.0
outcome==1.3.0.post0
packaging==25.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pylint==3.3.8
PySocks==1.7.1
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
ruff==0.13.3
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
tomlkit==0.13.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
